azure_metrics/requirements.txt
# integration pip requirements
azure-monitor==0.3.1